Application running but very buggy

+++ src/main/kotlin/com/writinggame/model/GameSettings.kt
@@ -1,3 +1,3 @@
 package com.writinggame.model
 
-class GameSettings(val roundTimeMinutes: Long? = 10, val minWordsPerMessage: Int? = null, val maxWordsPerMessage: Int? = null)
+class GameSettings(val roundTimeMinutes: Int? = 10, val minWordsPerMessage: Int? = null, val maxWordsPerMessage: Int? = null)
